test(cypress): move sanity client definition outside of test
